Add link to display request. Working but throwing test error
