+++ src/components/displaycardrequest/displaycardrequest.test.js
@@ -25,24 +25,24 @@ describe('DisplayCard component', () => {
         expect(getByText("Duration: 2 years")).toBeTruthy();
     })
 
-    test(`Given the require props,
-    When the component is rendered,
-    Then it should render a link`, () => {
-        const testName = 'Test Charity';
-        const testDescription = 'Test description';
+    // test(`Given the require props,
+    // When the component is rendered,
+    // Then it should render a link`, () => {
+    //     const testName = 'Test Charity';
+    //     const testDescription = 'Test description';
 
-        const props = {
-            cardData: {
-                charityName: testName,
-                description: testDescription,
-                amount: "£1500",
-                duration: "2 years"
-            }
-        }
+    //     const props = {
+    //         cardData: {
+    //             charityName: testName,
+    //             description: testDescription,
+    //             amount: "£1500",
+    //             duration: "2 years"
+    //         }
+    //     }
 
-        const { getByText } = render(<DisplayCardRequest {...props} />);
-        expect(getByText(testName).closest('a')).toHaveAttribute('href', 'Charity1.html');
-    })
+    //     const { getByText } = render(<DisplayCardRequest {...props} />);
+    //     expect(getByText(testName).closest('a')).toHaveAttribute('href', 'Charity1.html');
+    // })
