+++ src/components/displayrequest/DisplayRequest.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import Container from 'react-bootstrap/Container';
+
+    function DisplayRequest() {
+
+    return (
+        <Container>
+            <h3>
+                Hello Request
+            </h3>
+        </Container>
+
+    );
+
+}
+
+
+
+export default DisplayRequest;
